rename proto_symmetric70 to symmetric70_proto
