+++ keyboards/handwired/symmetric70_proto/config.h
@@ -24,7 +24,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define PRODUCT_ID   0x0000
 #define DEVICE_VER   0x0001
 #define MANUFACTURER mtei
-#define PRODUCT      proto_symmetric70
+#define PRODUCT      symmetric70_proto
 
 /* key matrix size */
 #define MATRIX_ROWS 5
