+++ keyboards/handwired/symmetric70_proto/info.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "proto_symmetric70",
+    "keyboard_name": "symmetric70_proto",
     "url": "",
     "maintainer": "mtei",
     "width": 16,
