+++ keyboards/handwired/symmetric70_proto/keymaps/default_mac/readme.md
@@ -1,4 +1,4 @@
-# The default_mac keymap for proto_symmetric70
+# The default_mac keymap for symmetric70_proto
 
 ```
   ESC   1  2  3  4  5                  6  7  8  9  0   BS
